rename tests for better readability, apply changes according to recommendation for the PR
