+++ selenium/che-selenium-test/src/test/resources/suites/CheSuite.xml
@@ -47,8 +47,8 @@
           <class name="org.eclipse.che.selenium.assistant.CheckFindActionFeatureInCheTest"/>
           <class name="org.eclipse.che.selenium.assistant.KeyBindingsTest"/>
           <class name="org.eclipse.che.selenium.assistant.OrganizeImportsTest"/>
-          <class name="org.eclipse.che.selenium.languageserver.CheckMainFeatureForCSharpLanguageTest"/>
-          <class name="org.eclipse.che.selenium.languageserver.PhpLanguageServerTest"/>
+          <class name="org.eclipse.che.selenium.languageserver.CSharpFileEditingTest"/>
+          <class name="org.eclipse.che.selenium.languageserver.PhpFileEditingTest"/>
           <class name="org.eclipse.che.selenium.dashboard.CreateAndDeleteProjectsTest"/>
           <class name="org.eclipse.che.selenium.dashboard.DeleteRunningWorkspaceTest"/>
           <class name="org.eclipse.che.selenium.dashboard.DeleteStoppingWorkspaceTest"/>
